Add 16 tracks, 32-step sequencer, Sampler one-shot machine, per-track mute
Digitakt-direction groundwork (keeps all Torso/S-4 features intact):

Scale to 16 tracks:
- kNumTracks 4 -> 16. Fixed ModEngine ADSR trig mask (uint32 -> uint64);
  packing track*slots overflowed past track 8 at 16 tracks.

Step sequencer (portable, RT-safe, JUCE-free):
- New StepSequencer (clock->step boundary converter), Pattern + PatternManager
  bank (separate from the Scene bank), wired into Engine on the audio thread.
- 32 steps shown as 16 buttons over 2 pages; unit-tested for ordering, wrap,
  block-size independence, and custom length.

Sampler machine + transport:
- New per-track MaterialMachine param: Torso (free-run tape/granular) vs
  Sampler (Digitakt voice). Default = Sampler; nothing plays on open until
  sequenced. Sequencer drives only Sampler tracks.
- Sampler = one-shot: plays the loop region once per trig (tape non-loop),
  granular silenced. Click-free via a crossfaded retrigger seek plus a short
  one-shot fade-in/out envelope in TapePlayer.
- Master PLAY ALL transport: starts the sequencer and launches Torso tracks.
- Per-track mute mode (red); removed the 16 per-track PLAY buttons.

+++ Source/modulation/ModEngine.cpp
@@ -59,7 +59,7 @@ void ModEngine::triggerAdsr (int track, int slot)
 {
     if (track < 0 || track >= kNumTracks || slot < 0 || slot >= kModSlotsPerTrack)
         return;
-    const uint32_t bit = 1u << static_cast<uint32_t> (track * kModSlotsPerTrack + slot);
+    const uint64_t bit = 1ull << static_cast<uint64_t> (track * kModSlotsPerTrack + slot);
     pendingAdsrTrig_.fetch_or (bit, std::memory_order_relaxed);
 }
 
@@ -140,12 +140,12 @@ void ModEngine::apply (ParameterState& params, float inputEnv01, int numSamples,
                                  : 0.0;
     const double beatAtBlockEnd = beatAtBlockStart + beatDelta;
 
-    const uint32_t trig = pendingAdsrTrig_.exchange (0, std::memory_order_acq_rel);
+    const uint64_t trig = pendingAdsrTrig_.exchange (0, std::memory_order_acq_rel);
     for (int t = 0; t < kNumTracks; ++t)
     {
         for (int s = 0; s < kModSlotsPerTrack; ++s)
         {
-            if (trig & (1u << static_cast<uint32_t> (t * kModSlotsPerTrack + s)))
+            if (trig & (1ull << static_cast<uint64_t> (t * kModSlotsPerTrack + s)))
                 adsrs_[static_cast<size_t> (t)][static_cast<size_t> (s)].trigger();
         }
     }
